Auton should maybe work.
Still needs to be tested. Code for not being on the boiler side not
included, but we might be able to get away with the boiler code for the
other side too.

+++ FRC2017/src/org/usfirst/frc/team5066/controller2017/AutonControlScheme.java
@@ -183,5 +183,5 @@ public abstract class AutonControlScheme {
 	 * @return The first row of the array refers to the case number in autonPeriodic,
 	 * the second row is for the time the motors will drive for, 0 if not needed for the case.
 	 */
-	public abstract double[] getSteps();
+	public abstract int[] getSteps();
 }
